requirements.txt
mysql-connector-python
python-dotenv
opencv-python
pika
imutils
sklearn
joblib
